+++ EnhanceQoL/Settings/Profiles.lua
@@ -1,3 +1,4 @@
+-- luacheck: globals EnhanceQoLBagsDB
 local addonName, addon = ...
 
 local L = LibStub("AceLocale-3.0"):GetLocale(addonName)